PLAT-10342 Moved datafeed package to symphony/core/service
